+++ hoot-rnd/src/main/cpp/hoot/rnd/criterion/ImplicitTagEligibleCriterion.h
