+++ hoot-rnd/src/main/cpp/hoot/rnd/criterion/ImplicitTagEligiblePoiCriterion.cpp
