+++ hoot-rnd/src/main/cpp/hoot/rnd/criterion/ImplicitTagEligiblePoiPolyCriterion.cpp
