+++ hoot-rnd/src/main/cpp/hoot/rnd/io/ImplicitTagRulesSqliteReader.cpp
