+++ hoot-rnd/src/main/cpp/hoot/rnd/io/ImplicitTagRulesSqliteWriter.h
