+++ hoot-rnd/src/main/cpp/hoot/rnd/visitors/ImplicitTypeTaggerBase.h
@@ -29,7 +29,7 @@
 
 // hoot
 #include <hoot/core/elements/ElementVisitor.h>
-#include <hoot/core/io/ImplicitTagRulesSqliteReader.h>
+#include <hoot/rnd/io/ImplicitTagRulesSqliteReader.h>
 #include <hoot/core/util/Configurable.h>
 #include <hoot/core/language/ToEnglishTranslator.h>
