+++ hoot-rnd/src/test/cpp/hoot/rnd/visitors/ImplicitPoiTypeTaggerTest.cpp
@@ -27,7 +27,7 @@
 
 // Hoot
 #include <hoot/core/TestUtils.h>
-#include <hoot/core/io/ImplicitTagRulesSqliteWriter.h>
+#include <hoot/rnd/io/ImplicitTagRulesSqliteWriter.h>
 #include <hoot/core/io/OsmJsonReader.h>
 #include <hoot/core/language/ToEnglishDictionaryTranslator.h>
 #include <hoot/core/util/Log.h>
